requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
DataRecorder==3.6.2
download==0.3.5
DownloadKit==2.0.7
DrissionPage==4.1.0.12
easyocr==1.7.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
hupper==1.12.1
idna==3.10
imageio==2.36.1
iniconfig==2.0.0
instagrapi==2.1.3
Jinja2==3.1.4
lazy_loader==0.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
mycdp==1.1.0
networkx==3.4.2
ninja==1.11.1.2
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
PasteDeploy==3.1.0
pdbp==1.6.1
pillow==11.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.3.6
pluggy==1.5.0
plyer==2.1.0
propcache==0.2.1
psutil==6.1.0
py==1.11.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
PyMsgBox==1.0.9
pynose==1.5.3
pyotp==2.9.0
pyperclip==1.9.0
pypiwin32==223
pyramid==2.0.2
pyramid_jinja2==2.10.1
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.4
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-bidi==0.6.3
python-dotenv==1.0.1
pytweening==1.2.0
pywin32==308
PyYAML==6.0.2
random-user-agent==1.0.1
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
sbvirtualdisplay==1.3.0
scikit-image==0.24.0
scipy==1.14.1
selenium==4.27.1
selenium-stealth==1.0.6
setuptools==75.6.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sympy==1.13.1
tabcompleter==1.4.0
tifffile==2024.9.20
tldextract==5.1.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
translationstring==1.4
trio==0.27.0
trio-websocket==0.11.1
ttkbootstrap==1.10.1
undetected-chromedriver==3.5.5
urllib3==2.2.3
venusian==3.1.1
WebOb==1.8.9
websocket-client==1.8.0
websockets==14.1
wheel==0.45.1
win10toast==0.9
wsproto==1.2.0
yarl==1.18.3
zenrows==1.4.0
zope.deprecation==5.0
zope.interface==7.2
